Cleaned up header and #include<>

+++ src/Main.cpp
@@ -82,7 +82,7 @@ size_t Main::numTasks() const
 }
 
 /////////////////////////////////////////////////////////////////////////////
-const Task* Main::getTask(unsigned int n) const
+const Task* Main::getTask(size_t n) const
 {
     return task[n];
 }
